+++ packages/kit/src/runtime/server/page/index.js
@@ -1,3 +1,4 @@
+import { devalue } from 'devalue';
 import { negotiate } from '../../../utils/http.js';
 import { render_response } from './render.js';
 import { respond_with_error } from './respond_with_error.js';
@@ -8,6 +9,7 @@ import { error, json } from '../../../exports/index.js';
 import { compact } from '../../../utils/array.js';
 import { normalize_error } from '../../../utils/error.js';
 import { load_data, load_server_data } from './load_data.js';
+import { DATA_SUFFIX } from '../../../constants.js';
 
 /**
  * @typedef {import('./types.js').Loaded} Loaded
@@ -48,8 +50,6 @@ export async function render_page(event, route, page, options, state, resolve_op
 			return handle_json_request(event, options, node.server);
 		}
 	}
-
-	const { fetcher, fetched, cookies } = create_fetch({ event, options, state, route });
 
 	try {
 		const nodes = await Promise.all([
@@ -102,7 +102,7 @@ export async function render_page(event, route, page, options, state, resolve_op
 		}
 
 		const should_prerender_data = nodes.some((node) => node?.server);
-		const data_pathname = `${event.url.pathname.replace(/\/$/, '')}/__data.json`;
+		const data_pathname = event.url.pathname.replace(/\/$/, '') + DATA_SUFFIX;
 
 		// it's crucial that we do this before returning the non-SSR response, otherwise
 		// SvelteKit will erroneously believe that the path has been prerendered,
@@ -111,18 +111,24 @@ export async function render_page(event, route, page, options, state, resolve_op
 		if (should_prerender) {
 			const mod = leaf_node.server;
 			if (mod && (mod.POST || mod.PUT || mod.DELETE || mod.PATCH)) {
-				throw new Error('Cannot prerender pages that have endpoints with mutative methods');
+				throw new Error('Cannot prerender pages that have mutative methods');
 			}
 		} else if (state.prerendering) {
 			// if the page isn't marked as prerenderable (or is explicitly
 			// marked NOT prerenderable, if `prerender.default` is `true`),
 			// then bail out at this point
-			if (!should_prerender) {
-				return new Response(undefined, {
-					status: 204
-				});
-			}
+			return new Response(undefined, {
+				status: 204
+			});
 		}
+
+		const { fetcher, fetched, cookies } = create_fetch({
+			event,
+			options,
+			state,
+			route,
+			prerender_default: should_prerender
+		});
 
 		if (get_option(nodes, 'ssr') === false) {
 			return await render_response({
@@ -166,7 +172,6 @@ export async function render_page(event, route, page, options, state, resolve_op
 					}
 
 					return await load_server_data({
-						dev: options.dev,
 						event,
 						state,
 						node,
@@ -231,12 +236,14 @@ export async function render_page(event, route, page, options, state, resolve_op
 
 					if (error instanceof Redirect) {
 						if (state.prerendering && should_prerender_data) {
+							const body = `window.__sveltekit_data = ${JSON.stringify({
+								type: 'redirect',
+								location: error.location
+							})}`;
+
 							state.prerendering.dependencies.set(data_pathname, {
-								response: new Response(undefined),
-								body: JSON.stringify({
-									type: 'redirect',
-									location: error.location
-								})
+								response: new Response(body),
+								body
 							});
 						}
 
@@ -294,12 +301,14 @@ export async function render_page(event, route, page, options, state, resolve_op
 		}
 
 		if (state.prerendering && should_prerender_data) {
+			const body = `window.__sveltekit_data = ${devalue({
+				type: 'data',
+				nodes: branch.map((branch_node) => branch_node?.server_data)
+			})}`;
+
 			state.prerendering.dependencies.set(data_pathname, {
-				response: new Response(undefined),
-				body: JSON.stringify({
-					type: 'data',
-					nodes: branch.map((branch_node) => branch_node?.server_data)
-				})
+				response: new Response(body),
+				body
 			});
 		}