+++ packages/kit/src/runtime/server/page/respond_with_error.js
@@ -35,7 +35,6 @@ export async function respond_with_error({ event, options, state, status, error,
 
 		if (ssr) {
 			const server_data_promise = load_server_data({
-				dev: options.dev,
 				event,
 				state,
 				node: default_layout,