+++ packages/kit/test/apps/options/test/test.js
@@ -163,10 +163,14 @@ test.describe('trailingSlash', () => {
 		expect(await r2.text()).toBe('hi');
 	});
 
-	test('can fetch data from page-endpoint', async ({ request, baseURL }) => {
-		const r = await request.get('/path-base/page-endpoint/__data.json');
-		expect(r.url()).toBe(`${baseURL}/path-base/page-endpoint/__data.json`);
-		expect(await r.json()).toEqual({
+	test('can fetch data from page-endpoint', async ({ request }) => {
+		const r = await request.get('/path-base/page-endpoint/__data.js');
+		const code = await r.text();
+
+		const window = {};
+		new Function('window', code)(window);
+
+		expect(window.__sveltekit_data).toEqual({
 			type: 'data',
 			nodes: [null, { type: 'data', data: { message: 'hi' }, uses: {} }]
 		});
@@ -184,7 +188,7 @@ test.describe('trailingSlash', () => {
 
 		/** @type {string[]} */
 		let requests = [];
-		page.on('request', (r) => requests.push(r.url()));
+		page.on('request', (r) => requests.push(new URL(r.url()).pathname));
 
 		// also wait for network processing to complete, see
 		// https://playwright.dev/docs/network#network-events
@@ -197,7 +201,7 @@ test.describe('trailingSlash', () => {
 			expect(requests.filter((req) => req.endsWith('.js')).length).toBeGreaterThan(0);
 		}
 
-		expect(requests.includes(`${baseURL}/path-base/prefetching/prefetched/__data.json`)).toBe(true);
+		expect(requests.includes(`/path-base/prefetching/prefetched/__data.js`)).toBe(true);
 
 		requests = [];
 		await app.goto('/path-base/prefetching/prefetched');