+++ packages/kit/types/internal.d.ts
@@ -16,7 +16,13 @@ import {
 	ServerInitOptions,
 	SSRManifest
 } from './index.js';
-import { HttpMethod, MaybePromise, RequestOptions, TrailingSlash } from './private.js';
+import {
+	HttpMethod,
+	MaybePromise,
+	PrerenderOption,
+	RequestOptions,
+	TrailingSlash
+} from './private.js';
 
 export interface ServerModule {
 	Server: typeof InternalServer;
@@ -133,7 +139,6 @@ export interface PageNode {
 
 export type PayloadScriptAttributes =
 	| { type: 'data'; url: string; body?: string }
-	| { type: 'server_data' }
 	| { type: 'validation_errors' };
 
 export interface PrerenderDependency {
@@ -268,7 +273,7 @@ export interface SSRNode {
 	shared: {
 		load?: Load;
 		hydrate?: boolean;
-		prerender?: boolean;
+		prerender?: PrerenderOption;
 		router?: boolean;
 		ssr?: boolean;
 	};
@@ -276,7 +281,7 @@ export interface SSRNode {
 	server: {
 		load?: ServerLoad;
 		hydrate?: boolean;
-		prerender?: boolean;
+		prerender?: PrerenderOption;
 		router?: boolean;
 		ssr?: boolean;
 		POST?: Action;
@@ -336,7 +341,9 @@ export interface PageNodeIndexes {
 	leaf: number;
 }
 
-export type SSREndpoint = Partial<Record<HttpMethod, RequestHandler>>;
+export type SSREndpoint = Partial<Record<HttpMethod, RequestHandler>> & {
+	prerender?: PrerenderOption;
+};
 
 export interface SSRRoute {
 	id: string;
@@ -355,6 +362,11 @@ export interface SSRState {
 	initiator?: SSRRoute | SSRErrorPage;
 	platform?: any;
 	prerendering?: PrerenderOptions;
+	/**
+	 * When fetching data from a +server.js endpoint in `load`, the page's
+	 * prerender option is inherited by the endpoint, unless overridden
+	 */
+	prerender_default?: PrerenderOption;
 }
 
 export type StrictBody = string | Uint8Array;